Added TerminalInterface and TerminalOutput to Terminal Component
