+++ package.json
@@ -21,16 +21,13 @@
     "@next/font": "13.0.4",
     "accept-language": "^3.0.18",
     "clsx": "^1.2.1",
-    "date-fns": "2.29.3",
     "lodash": "^4.17.21",
     "negotiator": "^0.6.3",
     "next": "13.0.4",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-individual-character-input-boxes": "^1.2.1",
     "react-modal": "^3.16.1",
     "react-responsive": "^9.0.2",
-    "react-terminal-ui": "^1.0.0",
     "react-transition-group": "^4.4.5",
     "use-local-storage-state": "^18.1.2"
   },
