--- ui/lesson/TerminalOutput.tsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-const TerminalOutput = ({ children }) => {
-  return <div className="react-terminal-line">{children}</div>
-}
-
-export { TerminalOutput }
